streamlit/requirements.txt
langchain==0.0.234
moviepy==1.0.3
openai==0.27.1
pytube==15.0.0
requests==2.31.0
chromadb==0.4.11
pyttsx3
tiktoken
elevenlabs
fairseq
g2p-en
Pillow
requests
soundfile
torch
youtube-transcript-api
googletrans==4.0.0-rc1
gTTS
nltk
streamlit
tenacity
openai
google-api-python-client
pytube
transformers
python-dotenv
google-auth-httplib2
google-auth-oauthlib
google-cloud-videointelligence
opencv-python
google-auth
ngrok
pandas
tabulate
dash
dash-bootstrap-components
plotly
matplotlib
timedelta
faiss-cpu
textblob
vaderSentiment
